fix np.float to np.float32

+++ lib/test_utils.py
@@ -153,9 +153,9 @@ def reconstruct_3D(depth, f):
         y = (v - cv) * depth / f
         z = depth
 
-    x = np.reshape(x, (width * height, 1)).astype(np.float)
-    y = np.reshape(y, (width * height, 1)).astype(np.float)
-    z = np.reshape(z, (width * height, 1)).astype(np.float)
+    x = np.reshape(x, (width * height, 1)).astype(np.float32)
+    y = np.reshape(y, (width * height, 1)).astype(np.float32)
+    z = np.reshape(z, (width * height, 1)).astype(np.float32)
     pcd = np.concatenate((x, y, z), axis=1)
     pcd = pcd.astype(np.int)
     return pcd
